refactored MpasGrid class definition to separate file.
